fixed the issue with weird stuff happening in the overview template.
turned out to tbe the fault of a wrongly closed div in the list template...

Next check if maybe I can actually use the buttons on the accordeon or the tabs to fetch the tables with htmx instead of fetching them all on reveal trigger. THis seems like a more efficient approach, Currently they are all fetcher at the same time.